feat: add display-box component
Add a new file `display-box.js` which defines a custom element `display-box`. The `display-box` element is a simple box that displays text. It has a shadow DOM and a template that contains a styled div element. The text content of the div element can be set using the `displayText` property.

feat: add index.html

Add a new file `index.html` which includes the component scripts for `user-input.js` and `display-box.js`. The HTML file also includes the `user-input` and `display-box` custom elements. When the DOM is loaded, an event listener is added to the `user-input` component. When the `userInputSubmit` event is fired, the text entered by the user is set as the `displayText` property of the `display-box` component.

feat(user-input.js): add UserInput component with input field and submit button